Fix Integer reference comparison bug in reconstructPath
Use intValue() instead of != to avoid reference comparison failure
for node indices >= 128 (outside Integer cache range).

+++ src/main/java/com/williamfiset/algorithms/graphtheory/BreadthFirstSearchAdjacencyList.java
@@ -63,7 +63,7 @@ public class BreadthFirstSearchAdjacencyList {
     for (Integer at = end; at != null; at = prev[at]) {
       path.addFirst(at);
     }
-    if (path.isEmpty() || path.getFirst() != start) {
+    if (path.isEmpty() || path.getFirst().intValue() != start) {
       return List.of();
     }
     return path;
